+++ yarpctest/fake_peer.go
@@ -27,7 +27,9 @@ import (
 
 // FakePeer is a fake peer with an identifier.
 type FakePeer struct {
-	id hostport.PeerIdentifier
+	id          hostport.PeerIdentifier
+	subscribers []peer.Subscriber
+	status      peer.Status
 }
 
 // Identifier returns the fake peer identifier.
@@ -37,16 +39,31 @@ func (p *FakePeer) Identifier() string {
 
 // Status returns the fake peer status.
 func (p *FakePeer) Status() peer.Status {
-	return peer.Status{
-		ConnectionStatus:    peer.Available,
-		PendingRequestCount: 0,
-	}
+	return p.status
 }
 
-// StartRequest does nothing.
+// StartRequest increments pending request count.
 func (p *FakePeer) StartRequest() {
+	p.status.PendingRequestCount++
 }
 
-// EndRequest does nothing.
+// EndRequest decrements pending request count.
 func (p *FakePeer) EndRequest() {
+	p.status.PendingRequestCount--
+}
+
+func (p *FakePeer) simulateConnect() {
+	p.status.ConnectionStatus = peer.Available
+	p.broadcast()
+}
+
+func (p *FakePeer) simulateDisconnect() {
+	p.status.ConnectionStatus = peer.Unavailable
+	p.broadcast()
+}
+
+func (p *FakePeer) broadcast() {
+	for _, sub := range p.subscribers {
+		sub.NotifyStatusChanged(p.id)
+	}
 }